+++ docs/conf.py
@@ -2,8 +2,6 @@ import datetime
 import os
 import textwrap
 
-import yaml
-
 # Configuration for the Sphinx documentation builder.
 # All configuration specific to your project should be done in this file.
 #
@@ -13,150 +11,91 @@ import yaml
 # A complete list of built-in Sphinx configuration values:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html
 #
-# Our Starter Pack uses the custom Canonical Sphinx extension
-# to keep all documentation based on it consistent and on brand:
+# The Sphinx Stack uses the Canonical Sphinx theme to keep all documentation consistent
+# and on brand:
 # https://github.com/canonical/canonical-sphinx
-
 
 #######################
 # Project information #
 #######################
 
 # Project name
-#
-# TODO: Update with the official name of your project or product
+# TODO: Update with the official name of your project or product (e.g., "Ubuntu Server")
+project = "Project"
 
-project = "Documentation Starter Pack"
+# Author name; used in the default copyright statement in the page footer
 author = "Canonical Ltd."
 
-# The year in the copyright statement defaults to the current year, so
-# individual document versions show when they were built.
-# TODO: If the date must be a range, like in a software license, replace
-# 2026 with the starting year of development and use:
-#
-# copyright = f"2026-{datetime.date.today().year}"
-
+# The year in the copyright statement
 copyright = f"{datetime.date.today().year}"
 
-# Sidebar documentation title; best kept reasonably short
-#
-# TODO: To include a version number, add it here (hardcoded or automated).
-#
-# TODO: To disable the title, set to an empty string.
-
+# Sidebar documentation title
+# To disable the title, set it to an empty string.
 html_title = project + " documentation"
 
-
 # Documentation website URL
-#
-# TODO: Update with the official URL of your docs or leave empty if unsure.
-#
-# NOTE: The Open Graph Protocol (OGP) enhances page display in a social graph
-#       and is used by social media platforms; see https://ogp.me/
-
-ogp_site_url = "https://canonical-starter-pack.readthedocs-hosted.com/"
-
+ogp_site_url = os.environ.get("READTHEDOCS_CANONICAL_URL", "/")
 
 # Preview name of the documentation website
-#
-# TODO: To use a different name for the project in previews, update as needed.
-
+# TODO: To use a different name for the project in previews, update the next line.
 ogp_site_name = project
 
-
 # Preview image URL
-#
-# TODO: To customise the preview image, update as needed.
-
+# TODO: To customise the preview image, update the next line.
 ogp_image = "https://assets.ubuntu.com/v1/cc828679-docs_illustration.svg"
 
-
 # Product favicon; shown in bookmarks, browser tabs, etc.
-
-# TODO: To customise the favicon, uncomment and update as needed.
-
-# html_favicon = '_static/favicon.png'
-
+# TODO: To customise the favicon, uncomment and update the next line.
+# html_favicon = ".sphinx/_static/favicon.png"
 
 # Dictionary of values to pass into the Sphinx context for all pages:
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#confval-html_context
-
 html_context = {
     # Product page URL; can be different from product docs URL
-    #
-    # TODO: Change to your product website URL,
-    #       dropping the 'https://' prefix, e.g. 'ubuntu.com/lxd'.
-    #
-    # TODO: If there's no such website,
-    #       remove the {{ product_page }} link from the page header template
-    #       (usually _templates/header.html; also, see README.rst).
-    "product_page": "documentation.ubuntu.com",
+    # TODO: Change to your product website URL, dropping the 'https://' prefix (e.g.,
+    #       'ubuntu.com/lxd'). If there's no such website, remove the {{ product_page }}
+    #       link from the _templates/header.html file.
+    "product_page": "",
     # Product tag image; the orange part of your logo, shown in the page header
-    #
     # TODO: To add a tag image, uncomment and update as needed.
     # 'product_tag': '_static/tag.png',
     # Your Discourse instance URL
-    #
     # TODO: Change to your Discourse instance URL or leave empty.
-    #
-    # NOTE: If set, adding ':discourse: 123' to an .rst file
-    #       will add a link to Discourse topic 123 at the bottom of the page.
-    "discourse": "https://discourse.ubuntu.com",
+    "discourse": "",
     # Your Mattermost channel URL
-    #
     # TODO: Change to your Mattermost channel URL or leave empty.
-    "mattermost": "https://chat.canonical.com/canonical/channels/documentation",
+    "mattermost": "",
     # Your Matrix channel URL
-    #
     # TODO: Change to your Matrix channel URL or leave empty.
-    "matrix": "https://matrix.to/#/#documentation:ubuntu.com",
-    # Your documentation GitHub repository URL
-    #
+    "matrix": "",
+    # Your documentation GitHub repository URL If set, links for viewing the
+    # documentation source files and creating GitHub issues are added at the bottom of
+    # each page.
     # TODO: Change to your documentation GitHub repository URL or leave empty.
-    #
-    # NOTE: If set, links for viewing the documentation source files
-    #       and creating GitHub issues are added at the bottom of each page.
-    "github_url": "https://github.com/canonical/sphinx-docs-starter-pack",
+    "github_url": "",
     # Docs branch in the repo; used in links for viewing the source files
-    #
-    # TODO: To customise the branch, uncomment and update as needed.
     "repo_default_branch": "main",
     # Docs location in the repo; used in links for viewing the source files
-    #
-    # TODO: To customise the directory, uncomment and update as needed.
     "repo_folder": "/docs/",
     # TODO: To enable or disable the Previous / Next buttons at the bottom of pages
     # Valid options: none, prev, next, both
-    # "sequential_nav": "both",
+    # "sequential_nav": "",
     # TODO: To enable listing contributors on individual pages, set to True
     "display_contributors": False,
     # Required for feedback button
     "github_issues": "enabled",
-    # Inherit the author value
+    # Passes the top-level 'author' value to the theme
     "author": author,
-    # The Starter Pack uses CC-BY-SA as the license
-    #
-    # TODO: If your docs need another license, specify it instead of 'CC-BY-SA'.
-    # For the name, we recommend using the standard shorthand identifier from
-    # https://spdx.org/licenses
-    #
-    # For the URL, link directly to the license statement, typically found on
-    # the product's home page or in its GitHub project.
-    #
-    # TODO: If your documentation is a part of the code repository of your project,
-    #       it inherits the code license instead; specify it instead of 'CC-BY-SA'.
+    # Documentation license information
     "license": {
-        "name": "CC-BY-SA-3.0",
-        "url": "https://github.com/canonical/sphinx-docs-starter-pack/blob/main/LICENSE",
+        # TODO: Specify your project's license.
+        # For the name, we recommend using the standard shorthand identifier from
+        # https://spdx.org/licenses
+        "name": "",
+        # TODO: Link directly to your project's license statement.
+        "url": "",
     },
 }
-
-html_extra_path = []
-
-# Allow opt-in build of the OpenAPI "Hello" example so docs stay clean by default.
-if os.getenv("OPENAPI", ""):
-    tags.add("openapi")
-    html_extra_path.append("how-to/assets/openapi.yaml")
 
 # TODO: To enable the edit button on pages, uncomment and change the link to a
 # public repository on GitHub or Launchpad. Any of the following link domains
@@ -166,14 +105,12 @@ if os.getenv("OPENAPI", ""):
 # - https://git.launchpad.net/example
 #
 # html_theme_options = {
-# 'source_edit_link': 'https://github.com/canonical/sphinx-docs-starter-pack',
+# 'source_edit_link': 'https://github.com/canonical/sphinx-stack',
 # }
 
-# Project slug; see https://meta.discourse.org/t/what-is-category-slug/87897
-#
-# TODO: If your documentation is hosted on https://docs.ubuntu.com/,
-#       uncomment and update as needed.
-
+# Project slug
+# TODO: If your documentation is hosted on https://documentation.ubuntu.com/,
+#       uncomment and set to the RTD slug.
 # slug = ''
 
 #######################
@@ -181,27 +118,22 @@ if os.getenv("OPENAPI", ""):
 #######################
 
 # Use RTD canonical URL to ensure duplicate pages have a specific canonical URL
-
 html_baseurl = os.environ.get("READTHEDOCS_CANONICAL_URL", "/")
 
 # sphinx-sitemap uses html_baseurl to generate the full URL for each page:
-
 sitemap_url_scheme = "{link}"
 
 # Include `lastmod` dates in the sitemap:
-
 sitemap_show_lastmod = True
 
-# Exclude generated pages from the sitemap:
-
+# TODO: Exclude pages that aren't user-facing from the sitemap (e.g., module pages
+# generated by autodoc).
+# Pages excluded from the sitemap:
 sitemap_excludes = [
     "404/",
     "genindex/",
     "search/",
 ]
-
-# TODO: Add more pages to sitemap_excludes if needed. Wildcards are supported.
-#       For example, to exclude module pages generated by autodoc, add '_modules/*'.
 
 ################################
 # Template and asset locations #
@@ -209,7 +141,6 @@ sitemap_excludes = [
 
 # html_static_path = ["_static"]
 # templates_path = ["_templates"]
-
 
 #############
 # Redirects #
@@ -237,8 +168,8 @@ rediraffe_dir_only = True
 # <first sentence of home page>".
 llms_txt_description = textwrap.dedent(
     """\
-    This is the documentation for the Sphinx Starter Pack, a template repository
-    that helps you set up, build, and publish Sphinx documentation.
+    This is the documentation for the Sphinx Stack, a template repository that helps you
+    set up, build, and publish Sphinx documentation.
     """
 )
 
@@ -251,9 +182,6 @@ if os.environ.get("READTHEDOCS"):
 ###########################
 
 # A regex list of URLs that are ignored by 'make linkcheck'
-#
-# TODO: Remove or adjust the ACME entry after you update the contributing guide
-
 linkcheck_ignore = [
     "http://127.0.0.1:8000",
     "https://github.com",
@@ -263,13 +191,14 @@ linkcheck_ignore = [
     r"https://.*\.sourceforge\.(net|io)/.*",
 ]
 
-
 # A regex list of URLs where anchors are ignored by 'make linkcheck'
-
 linkcheck_anchors_ignore_for_url = [r"https://github\.com/.*"]
 
-# give linkcheck multiple tries on failure
+# How long the link checker will wait for a response for each request
+# TODO: Decrease to improve run time or increase if links frequently time out.
 # linkcheck_timeout = 30
+
+# Give linkcheck multiple tries on failure
 linkcheck_retries = 3
 
 ########################
@@ -278,18 +207,14 @@ linkcheck_retries = 3
 
 # Custom MyST syntax extensions; see
 # https://myst-parser.readthedocs.io/en/latest/syntax/optional.html
-#
 # NOTE: By default, the following MyST extensions are enabled:
-#       substitution, deflist, linkify
-
+#   - substitution
+#   - deflist
+#   - linkify
 # myst_enable_extensions = set()
-
 
 # Custom Sphinx extensions; see
 # https://www.sphinx-doc.org/en/master/usage/extensions/index.html
-
-# NOTE: The canonical_sphinx extension is required for the Starter Pack.
-
 extensions = [
     "canonical_sphinx",
     "notfound.extension",
@@ -315,65 +240,42 @@ extensions = [
 ]
 
 # Excludes files or directories from processing
-
 exclude_patterns = [
     "doc-cheat-sheet*",
     ".venv*",
 ]
 
-# Adds custom CSS files, located under 'html_static_path'
-
+# Adds custom CSS files, located remotely or in 'html_static_path'.
 # html_css_files = [
 #     "https://assets.ubuntu.com/v1/d86746ef-cookie_banner.css",
 # ]
 
-
-# Adds custom JavaScript files, located under 'html_static_path'
-
+# Adds custom JavaScript files, located remotely or in 'html_static_path'.
 # html_js_files = [
 #     "https://assets.ubuntu.com/v1/287a5e8f-bundle.js",
 # ]
 
-
-# Specifies a reST snippet to be appended to each .rst file
-# If you have many entries, consider creating a reuse/substitutions.txt file
-# and loading it here instead.
+# Appends extra markup to the end of every document written in reST
 # rst_epilog = """
-#     .. include:: reuse/substitutions.txt
-#     """
-rst_epilog = """
-.. |RST| replace:: :abbr:`reST (reStructuredText)`
-.. |version_number| replace:: 0.1.0
-.. |rest_text| replace:: *Multi-line* text
-                         that uses basic **markup**.
-.. |site_link| replace:: Website link
-.. _site_link: https://example.com
-"""
+# """
 
 # Feedback button at the top; enabled by default
-#
-# TODO: To disable the button, uncomment this.
-
+# TODO: Disable the button if your project is unsuitable for public feedback.
 # disable_feedback_button = True
 
-
 # Your manpage URL
-#
 # TODO: To enable manpage links, uncomment and replace {codename} with required
 #       release, preferably an LTS release (e.g. noble). Do *not* substitute
 #       {section} or {page}; these will be replaced by sphinx at build time
 #
 # NOTE: If set, adding ':manpage:' to an .rst file
 #       adds a link to the corresponding man section at the bottom of the page.
-
 # manpages_url = 'https://manpages.ubuntu.com/manpages/{codename}/en/' + \
 #     'man{section}/{page}.{section}.html'
-
 
 # Specifies a reST snippet to be prepended to each .rst file
 # This defines a :center: role that centers table cell content.
 # This defines a :h2: role that styles content for use with PDF generation.
-
 rst_prolog = """
 .. role:: center
    :class: align-center
@@ -385,30 +287,8 @@ rst_prolog = """
     :class: vale-ignore
 """
 
-# Workaround for https://github.com/canonical/canonical-sphinx/issues/34
-
-if "discourse_prefix" not in html_context and "discourse" in html_context:
-    html_context["discourse_prefix"] = f"{html_context['discourse']}/t/"
-
-# Workaround for substitutions.yaml
-# If the user has a reuse/substitutions.yaml file, load from there.
-# Otherwise, use the manual definitions below.
-
-if os.path.exists("./reuse/substitutions.yaml"):
-    with open("./reuse/substitutions.yaml", "r") as fd:
-        myst_substitutions = yaml.safe_load(fd.read())
-else:
-    myst_substitutions = {
-        "version_number": "0.1.0",
-        "formatted_text": "*Multi-line* text\n that uses basic **markup**.",
-        "site_link": "[Website link](https://example.com)"
-    }
-
-# Add configuration for intersphinx mapping
-# Map only the Sphinx documentation sets that you need to link to from your docs set.
-intersphinx_mapping = {
-    "sphinxcontrib-mermaid": (
-        "https://sphinxcontrib-mermaid-demo.readthedocs.io/en/latest",
-        None,
-    )
-}
+# Configuration for Intersphinx projects
+#
+# intersphinx_mapping = {
+#     "snap": ("https://snapcraft.io/docs/", None),
+# }